add helpers file, pull functions out of App.js
